+++ xuberservice/src/main/java/com/gox/xuberservice/model/XuperCheckRequest.kt
@@ -154,7 +154,7 @@ data class XuperCheckRequest(
                     val discount: Double? = 0.0,
                     val discount_percent: Double? = 0.0,
                     val distance: Double? = 0.0,
-                    val extra_charges: Int = 0,
+                    val extra_charges: Double = 0.0,
                     val extra_charges_notes: String? = "",
                     val fixed: Double? = 0.0,
                     val fleet: Double? = 0.0,
